requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
Flask==1.0.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
parsel==1.4.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.6.1
pyOpenSSL==18.0.0
PyYAML==3.12
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
urllib3==1.23
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.5.0
